Add cli Property to append value to package id
In the scenario where multiple branches (for different environments, like
master, dev, staging) have their packages built using octopack, and they
are published to the same nuget feed (e.g. they are all being pushed to
Teamcity's nuget feed), the packages will end up having the same id, which
will cause Octopus to pull the wrong package.

This property allow the user to pass in some arbirtrary value, the
environment name for example, to the msbuild task so that it can
differentiate the packages being generated.