Improves stability of results across test runs
As of Visual Studio 2019, you must now add a Test Adapter NuGet package
to unit test projects so that the unit tests can be discovered and run.
This commit adds the NUnit 2 test adapter NuGet package.

Also, the unit tests project was updated to include the ProjectTypeGuids
that let Visual Studio know the project is a test project and provide
the Run Tests option in the context menu of the project.

As an NCrunch user, NCrunch can greatly increase stability of unit tests
across test runs. This commit contains some changes that help improve
the stability of test results across test runs.

To improve stability across test runs, the post-build event on the
OctoPack.Tasks project was replaced with a custom target in the
OctoPack.Tests project that executes as the last target that is run as
part of a build of that project.

In addition, the OctoPack.Tests project did not directly reference the
OctoPack.Tests.SampleGitVersionAssembly or
OctoPack.Tests.NonGitVersionAssembly projects, but instead, used
complicated assembly locating logic which is fragile and simply not
usable when executing the AssemblyExtensionsTasks tests by NCrunch.
Instead, directly reference those assemblies. When configuring NCrunch
with
<CopyReferenceAssembliesToWorkspace>true</CopyReferenceAssembliesToWorkspace>,
"It just Works". (NCrunch also needs to be configured to include
additional files, such as the ..\..\build\**.* and ..\Samples\**.*
files.)

Lastly, I also added the OctoPack.Tests.SampleExeDependency project to
the solution. It was there--just never added to the solution.

With these changes in place, the tests are running locally and I can
safely make changes to the project.