Fixes OctoPack packages including extra files
Back in January, 2018, PR 2878 at dotnet/msbuild was merged. (See
https://www.github.com/dotnet/msbuild/pull/2878). There was one simple
change, but it unknowingly caused OctoPack to start including extra
unnecessary files into OctoPack packages. We found it because we had
deployments blow up in Octopus Deploy due to file paths which exceeded
MAX_PATH length during deployment, specifically, during config file
transformation.

The PR had the result of, when copying files marked as CopyLocal, a
*.csproj.CopyComplete file is created. It's a marker for the build
process so that transitively referenced projects are built
appropriately. Near as far as I can tell, because the file is created
(via the Touch task) during this target, and subsequently added to the
@(FileWrites) item group, and because the path associated with
@(CopyUpToDateMarker) item group is an absolute path, the existing
exclusion in the OctoPack.targets file is not filtering out these files.
So they are being included in the OctoPack deployment packages. While
the files are 0-bytes in size and are harmless, they don't belong in the
package. They are not a deployable build artifact, and in my case,
caused deployment failures.

This wasn't caught in changes to OctoPack between Jan. 2018 and Jan.
2019 (the last time any change was made to this repo, and the date of
the most recent release) because the unit tests were never appropriately
updated to take into account newer versions of Visual Studio, namely
Visual Studio 2017 and Visual Studio 2019.

This commit updates the unit tests and the OctoPack.targets file so that
these useless *.csproj.CopyComplete files are not added to OctoPack
deployment packages.