Removes build-generated files from the repository
This commit removes files generated during the build of projects in the
Samples solution from the repository and adds a .gitignore file to the
Samples directory to further ensure these files do not get accidentally
re-added to the repository.

These files, if present, can cause the unit tests to fail since the unit
tests currently do not do a great job of cleaning up after themselves.

This change highlighted another assumption made in one of the tests,
that one of the sample project files was already copied to
*.csproj.teamcity. So this test was updated to explicitly copy the
required file, further improving test stability in different
environments across test runs.