Updates OctoPack.nuspec to use best-practice layout
This commit updates the OctoPack.nuspec file to use a more best-practice
layout, as can be seen at Nate McMaster's blog (MS):
https://natemcmaster.com/blog/2017/11/11/build-tools-in-nuget

This commit:

* Updates the build.cake script to remove the .\build\contents\<tfm>
folder creation and instead creates .\build\lib\<tfm> folders, as this
is currently the best-practice for creating NuGet packaged build tools
such that they report to Visual Studio the target frameworks for which
the tools are supported.

* Updates the NuSpec file accordingly, using a 'lib' folder in place of
the 'content' folder. Also, this NuGet package can only be installed
with a NuGet client having a minimum  version of 2.5 or greater, so this
was made explicit by adding the minClientVersion attribute to the
Nuspec's <metadata /> element.

It's really hard to write a "portable end-to-end test" that the package
installs/uninstalls using both packages.config-style and
<PackageRefenence />-style installation/uninstallation methods, since
I'm not likely to have a development NuGet feed setup in just the same
way as someone else. But I did test these changes locally for both
styles of installing the package, building the project with '/t:Build
/p:Configuration=Release /p:RunOctoPack=true' and verifying the package
contents, and then uninstalling the package.